+++ entreprise.go
@@ -441,7 +441,7 @@ func (e *Etablissements) getBatch(roles scope, username string) *pgx.Batch {
 		left join etablissement_follow f on f.siret = et.siret and f.active and f.username = $3
 		where et.siren = any($4) 
 		and coalesce(s.libelle_liste, $5) = $5
-		order by ef.effectif desc;`,
+		order by ef.effectif desc, et.siret desc;`,
 		roles.zoneGeo(),
 		roles.zoneGeo(),
 		username,
@@ -472,11 +472,11 @@ func (e *Etablissements) loadEtablissements(rows *pgx.Rows) error {
 		if err != nil {
 			return err
 		}
-		cousins[score.Siret] = append(cousins[score.Siret], score)
+		cousins[score.Siret[0:9]] = append(cousins[score.Siret[0:9]], score)
 	}
 
 	for k, v := range cousins {
-		entreprise := e.Entreprises[k[0:9]]
+		entreprise := e.Entreprises[k]
 		entreprise.EtablissementsSummary = v
 		e.Entreprises[k[0:9]] = entreprise
 	}
